feat(release): boot-gate so a build that won't launch can't ship + recovery shim
Packaging is a separate surface from the source: a module can be left out of the
app.asar allowlist, or a bundled binary mispackaged, so the app builds and passes
static checks yet fails at startup. Building and file-inspection don't catch that;
this gate runs the packaged app and requires a real boot before it ships.

Static check - tests/require-coverage.lib.js + packaged-requires.test.js +
require-coverage.test.js assert every relative require() in a packaged electron JS
file is in build.files. A new `electron` job in ci.yml runs the electron tests on
every PR and feeds the required Test aggregator.

Boot smoke - build.yml launches the packaged app and polls /health/ready until
200 {"db":"ok"} (a live SELECT 1), failing fast if the process exits. Proves
electron -> Python API -> embedded Postgres -> a real query. Mac-only on PRs; all
three platforms at release, where it gates promotion to latest. CELERP_API_PORT pins
the API port for CI; unset, behavior is unchanged for users.

Recovery shim - main.js is now a thin wrapper that loads the real app (renamed to
app-main.js); if loading ever fails it shows a recovery message and exits instead of
failing silently. Transparent passthrough on a healthy build; app-main.js is in
build.files.
