+++ electron/restart.js
@@ -1,7 +1,7 @@
 // Copyright (c) 2026 Noah Severs
 // SPDX-License-Identifier: BUSL-1.1
 //
-// Pure functions extracted from main.js for unit testing.
+// Pure functions extracted from app-main.js for unit testing.
 // No Electron imports — safe to require in Jest.
 
 "use strict";
